==> step6_combine.ipynb <==
import numpy as np
import pandas as pd
from datetime import datetime, timedelta
import matplotlib.pyplot as plt

import os
import sys

###############################################################################
#   Housekeeping
###############################################################################
dir_input = os.path.join("./processing/")
dir_output = os.path.join("./output/")
if not os.path.exists(dir_output):
  os.makedirs(dir_output)

fname_drybulb = "Timeseries_step1_drybulb,raw_cloud,raw_relhum.csv"
fname_relhum = "Timeseries_step2_corrected_relhum.csv"
fname_wetbulb = "Timeseries_step3_wetbulb.csv"
fname_pres = "Timeseries_step4_wind,pressure.csv"
fname_wind = "Timeseries_step4_wind,pressure.csv"
fname_cloud = "Timeseries_step5_corrected_cloud.csv"


###############################################################################
#   Combine timeseries
###############################################################################
df_tmp = pd.read_csv(os.path.join(dir_input, fname_drybulb), usecols = ["Date", "Dry-bulb Temperature [F]"])
dt0 = df_tmp["Date"].iloc[0]
dt1 = df_tmp["Date"].iloc[-1]
index = pd.date_range(dt0, dt1, freq = timedelta(hours = 1))
df_tmp = df_tmp.set_index(index)

df_drybulb = df_tmp.drop("Date", axis = 1)

df_tmp = pd.read_csv(os.path.join(dir_input, fname_relhum), usecols = ["Date", "Relative Humidity [%]"])
dt0 = df_tmp["Date"].iloc[0]
dt1 = df_tmp["Date"].iloc[-1]
index = pd.date_range(dt0, dt1, freq = timedelta(hours = 1))
df_tmp = df_tmp.set_index(index)

df_relhum = df_tmp.drop("Date", axis = 1)

df_tmp = pd.read_csv(os.path.join(dir_input, fname_wetbulb), usecols = ["Date", "Wet-bulb Temperature [F]"])
dt0 = df_tmp["Date"].iloc[0]
dt1 = df_tmp["Date"].iloc[-1]
index = pd.date_range(dt0, dt1, freq = timedelta(hours = 1))
df_tmp = df_tmp.set_index(index)

df_wetbulb = df_tmp.drop("Date", axis = 1)

df_tmp = pd.read_csv(os.path.join(dir_input, fname_pres), usecols = ["Date", "Atm_Pres [inHg]"])
dt0 = df_tmp["Date"].iloc[0]
dt1 = df_tmp["Date"].iloc[-1]
index = pd.date_range(dt0, dt1, freq = timedelta(hours = 1))
df_tmp = df_tmp.set_index(index)

df_pres = df_tmp.drop("Date", axis = 1)

df_tmp = pd.read_csv(os.path.join(dir_input, fname_pres), usecols = ["Date", "Wind [mph]"])
dt0 = df_tmp["Date"].iloc[0]
dt1 = df_tmp["Date"].iloc[-1]
index = pd.date_range(dt0, dt1, freq = timedelta(hours = 1))
df_tmp = df_tmp.set_index(index)

df_wind = df_tmp.drop("Date", axis = 1)

df_tmp = pd.read_csv(os.path.join(dir_input, fname_cloud), usecols = ["Date", "Cloud Cover [fraction]"])
dt0 = df_tmp["Date"].iloc[0]
dt1 = df_tmp["Date"].iloc[-1]
index = pd.date_range(dt0, dt1, freq = timedelta(hours = 1))
df_tmp = df_tmp.set_index(index)

df_cloud = df_tmp.drop("Date", axis = 1)

df_all = pd.concat([df_drybulb, df_wetbulb, df_relhum, df_pres, df_cloud, df_wind], axis=1)

df_all.index.name = "Date"
df_all.to_csv(os.path.join(dir_output, "Timeseries_All.csv"),
                  float_format="%.3f",
                  sep=",",
                  na_rep="NaN")
